feat: pass Vault signer settings by VaultOpts
