+++ e2e/go.mod
@@ -3,7 +3,7 @@ module github.com/werf/trdl/e2e
 go 1.24.10
 
 require (
-	github.com/deckhouse/delivery-kit-sdk v1.2.1
+	github.com/deckhouse/delivery-kit-sdk v1.2.2-0.20260722202717-2aa49a87a836
 	github.com/hashicorp/go-hclog v1.6.3
 	github.com/hashicorp/vault/sdk v0.8.1
 	github.com/onsi/ginkgo/v2 v2.22.0
